requirements.txt
aenum==2.1.2
aiodns==2.0.0
aiofiles==0.4.0
aiohttp==3.5.4
aioitertools==0.4.0
aiosqlite==0.10.0
appnope==0.1.0
async-timeout==3.0.1
asyncio==3.4.3
attrs==19.1.0
backcall==0.1.0
cffi==1.12.3
chardet==3.0.4
ciso8601==2.1.1
colorama==0.4.1
dacite==0.0.26
decorator==4.4.0
idna==2.8
-e git+git@github.com:nickmflorin/instagram-attack.git@25fe8a1e27b21caeb467b7d010982499cdf5f7df#egg=instattack
ipdb==0.12
ipython==7.4.0
ipython-genutils==0.2.0
jedi==0.13.3
maxminddb==1.4.1
multidict==4.5.2
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
plumbum==1.6.7
progressbar2==3.39.3
prompt-toolkit==2.0.9
proxybroker==0.3.2
ptyprocess==0.6.0
pycares==3.0.0
pycparser==2.19
Pygments==2.3.1
PyPika==0.27.1
python-dotenv==0.10.1
python-utils==2.3.0
six==1.12.0
tortoise-orm==0.11.10
traitlets==4.3.2
wcwidth==0.1.7
yarl==1.3.0
